+++ src/targets/cpu/include/migraphx/cpu/gemm.hpp
@@ -11,7 +11,7 @@ namespace cpu {
 void migemm(
     const argument& c_arg, const argument& a_arg, const argument& b_arg, float alpha, float beta);
 void migemm(
-    const argument& c_arg, const argument& a_arg, const argument& b_arg, int8_t alpha, int8_t beta);
+    const argument& c_arg, const argument& a_arg, const argument& b_arg, int32_t alpha, int32_t beta);
 
 } // namespace cpu
 } // namespace MIGRAPHX_INLINE_NS
